requirements.txt
requests>=2.31.0
python-amazon-paapi>=5.0.0
requests
python-amazon-paapi
google-genai